refactor: Sync同期最適化(その1) 同期 id
- 同期物体用id 親子関係解決用 -> short型に変更

+++ Packages/tech.gyoku.fdmi.sync/Runtime/FDMiRelativeObjectSyncManager.cs
@@ -19,9 +19,9 @@ namespace tech.gyoku.FDMi.sync
             _rotation = Quaternion.identity;
             localRootRefPoint = this;
             ParentIndex = -1;
-            for (int i = 0; i < refPoints.Length; i++)
+            for (short i = 0; i < refPoints.Length; i++)
                 refPoints[i].index = i;
-            for (int i = 0; i < refPoints.Length; i++)
+            for (short i = 0; i < refPoints.Length; i++)
                 refPoints[i].initReferencePoint();
         }
         public Quaternion localRotation;
